refactor(rust-api): extract generation job handlers into generation.rs (sc-1636)
Move the image/video/vqa/interleave job-creation block out of lib.rs
into generation.rs: create_image_job, create_vqa_job, create_interleave_job,
create_video_job, their validators, and the recipe-preset payload-application
helpers. app_fallback/route_not_found stay at the root.

Same pattern: `use super::*;` + pub(crate). The cluster is pub(crate)
because parse_recipe_preset_resolution is also called from the
recipe-preset code outside the module. Behavior-preserving.
lib.rs 6,799 -> 6,405 lines.
